+++ python/lsst/donut/viz/aggregate_visit.py
@@ -139,6 +139,10 @@ class AggregateZernikeTablesTask(pipeBase.PipelineTask):
             raw_table["zk_CCS"] = np.atleast_2d(zernikes_merged[1:])
             raw_table["detector"] = det_meta["det_name"]
             raw_table["used"] = zernike_table["used"][1:]
+            if "extra_donut_id" in zernike_table.colnames:
+                raw_table["extra_donut_id"] = zernike_table["extra_donut_id"][1:]
+            if "intra_donut_id" in zernike_table.colnames:
+                raw_table["intra_donut_id"] = zernike_table["intra_donut_id"][1:]
             raw_tables.append(raw_table)
             avg_table = Table()
             avg_table["zk_CCS"] = np.atleast_2d(zernikes_merged[0])
@@ -899,6 +903,7 @@ class AggregateAOSVisitTableTask(pipeBase.PipelineTask):
             if single_sided:  # single-sided Zernike estimates
                 for k in avg_keys:
                     raw_table[k][w] = adt[k][wadt]
+                raw_table["donut_id"] = adt["donut_id"][wadt]
             else:  # double-sided Zernike estimates
                 wintra = adt[wadt]["focusZ"] == visit_fzmin
                 wextra = adt[wadt]["focusZ"] == visit_fzmax
@@ -917,6 +922,17 @@ class AggregateAOSVisitTableTask(pipeBase.PipelineTask):
                     raw_table[k + "_intra"][w] = adt[k][wadt][wintra]
                     raw_table[k + "_extra"][w] = adt[k][wadt][wextra]
 
+                # donut id can't be averaged like coordinates or centroids,
+                # so we process it separately
+                k = "donut_id"
+                if k in adt.colnames():  # safeguard against older data
+                    nrows = len(raw_table)
+                    dtype = adt[k].dtype
+                    if k + "_intra" not in raw_table.colnames:
+                        raw_table[k + "_intra"] = np.full(nrows, "", dtype=dtype)
+                        raw_table[k + "_extra"] = np.full(nrows, "", dtype=dtype)
+                    raw_table[k + "_intra"][w] = adt[k][wintra]
+                    raw_table[k + "_extra"][w] = adt[k][wextra]
         return pipeBase.Struct(raw=raw_table, avg=avg_table)
 
 
@@ -1006,6 +1022,17 @@ class AggregateAOSVisitTableCwfsTask(AggregateAOSVisitTableTask):
                 raw_table[k + "_intra"][w] = adt[k][wintra]
                 raw_table[k + "_extra"][w] = adt[k][wextra]
 
+            # donut id can't be averaged like coordinates or centroids,
+            # so we process it separately
+            k = "donut_id"
+            if k in adt.colnames:  # safeguard against older data
+                nrows = len(raw_table)
+                dtype = adt[k].dtype
+                if k + "_intra" not in raw_table.colnames:
+                    raw_table[k + "_intra"] = np.full(nrows, "", dtype=dtype)
+                    raw_table[k + "_extra"] = np.full(nrows, "", dtype=dtype)
+                raw_table[k + "_intra"][w] = adt[k][wintra]
+                raw_table[k + "_extra"][w] = adt[k][wextra]
         return pipeBase.Struct(raw=raw_table, avg=avg_table)